update new function to support terraform for cloud database services
- Replace BizFly to Bizfly, ProjectId to ProjectID
