+++ go.mod
@@ -2,4 +2,8 @@ module github.com/bizflycloud/gobizfly
 
 go 1.16
 
-require github.com/stretchr/testify v1.4.0
+require (
+	github.com/stretchr/testify v1.4.0
+	golang.org/x/lint v0.0.0-20210508222113-6edffad5e616 // indirect
+	golang.org/x/tools v0.14.0 // indirect
+)
